# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Puesto de Salud Bohórquez",
    "description": "API para gestionar pacientes, doctores, servicios, consultorios y citas.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
...
  "paths": {
...
    "/consultorios": {
      "get": {
        "tags": ["Consultorios"],
        "summary": "Listar consultorios",
        "responses": {
          "200": {
            "description": "Consultorios listados",
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Consultorio" } } }
            }
          }
        }
      },

      "post": {
        "tags": ["Consultorios"],
        "summary": "Registrar un consultorio",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsultorioCreate" } } }
        },
        "responses": {
          "201": { "description": "Consultorio creado" }
        }
      }
    },
    "/consultorios/{id}": {
      "put": {
        "tags": ["Consultorios"],
        "summary": "Actualizar un consultorio",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConsultorioCreate" } } }
        },
        "responses": {
          "200": { "description": "Consultorio actualizado" }
        }
      },

      "delete": {
        "tags": ["Consultorios"],
        "summary": "Eliminar un consultorio",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Consultorio eliminado" }
        }
      }
    },
    "/citas": {
      "get": {
        "tags": ["Citas"],
        "summary": "Listar todas las citas",
        "responses": {
          "200": {
            "description": "Lista de citas",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Cita" } }
              }
            }
          }
        }
      },

      "post": {
        "tags": ["Citas"],
        "summary": "Crear una cita",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CitaCreate" } } }
        },
        "responses": {
          "201": { "description": "Cita creada" },
          "400": { "description": "Datos inválidos" }
        }
      }
    },
    "/citas/{id}": {
      "put": {
        "tags": ["Citas"],
        "summary": "Actualizar una cita",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CitaCreate" } } }
        },
        "responses": {
          "200": { "description": "Cita actualizada" },
          "404": { "description": "Cita no encontrada" }
        }
      },

      "delete": {
        "tags": ["Citas"],
        "summary": "Eliminar una cita",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Cita eliminada" },
          "404": { "description": "Cita no encontrada" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Consultorio": {
        "type": "object",
        "properties": {
          "id_consultorio": { "type": "integer" },
          "nombre": { "type": "string" },
          "ubicacion": { "type": "string" },
          "tipo": { "type": "string" }
        }
      },
      "ConsultorioCreate": {
        "type": "object",
        "required": ["nombre", "ubicacion", "tipo"],
        "properties": {
          "nombre": { "type": "string" },
          "ubicacion": { "type": "string" },
          "tipo": { "type": "string" }
        }
      },
      "Cita": {
        "type": "object",
        "properties": {
          "id_cita": { "type": "integer" },
          "id_paciente": { "type": "integer" },
          "id_profesional": { "type": "integer" },
          "id_servicio": { "type": "integer" },
          "id_consultorio": { "type": "integer" },
          "fecha": { "type": "string" },
          "hora_inicio": { "type": "string" },
          "hora_fin": { "type": "string" },
          "estado": { "type": "string" },
          "notas": { "type": "string" }
        }
      },
      "CitaCreate": {
        "type": "object",
        "required": ["id_paciente", "id_profesional", "id_servicio", "id_consultorio", "fecha", "hora_inicio"],
        "properties": {
          "id_paciente": { "type": "integer" },
          "id_profesional": { "type": "integer" },
          "id_servicio": { "type": "integer" },
          "id_consultorio": { "type": "integer" },
          "fecha": { "type": "string" },
          "hora_inicio": { "type": "string" },
          "hora_fin": { "type": "string" },
          "estado": { "type": "string" },
          "notas": { "type": "string" }
        }
      }
    }
  }
}